# src/Services/AterStudio/openapi_admin.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "AterStudio | admin",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/Project/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/Project/service": {
      "post": {
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/Project/open": {
      "post": {
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/Project/setting/{id}": {
      "put": {
        "tags": [
          "Project"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfig"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfig"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectConfig"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionStatus": {
        "type": "integer",
        "x-enumNames": [
          "NotStarted",
          "InProgress",
          "Success",
          "Failed"
        ],
        "x-enumData": [
          {
            "name": "NotStarted",
            "value": 0,
            "description": "未执行"
          },
          {
            "name": "InProgress",
            "value": 1,
            "description": "执行中"
          },
          {
            "name": "Success",
            "value": 2,
            "description": "成功"
          },
          {
            "name": "Failed",
            "value": 3,
            "description": "失败"
          }
        ]
      },
...
      "ApiDocInfo": {
        "required": [
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "content": {
            "maxLength": 10000,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "nullable": true
          },
          "path": {
            "maxLength": 300,
            "type": "string"
          },
          "localPath": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        }
      },
...
      "ControllerType": {
        "type": "integer",
        "x-enumNames": [
          "Client",
          "Admin",
          "Both"
        ],
        "x-enumData": [
          {
            "name": "Client",
            "value": 0,
            "description": "用户端"
          },
          {
            "name": "Admin",
            "value": 1,
            "description": "管理端"
          },
          {
            "name": "Both",
            "value": 2,
            "description": "用户端和管理端"
          }
        ]
      },
...
      "GenAction": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 40,
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "entityPath": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "openApiPath": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "variables": {
            "type": "array",
            "items": { }
          },
          "sourceType": {
            "$ref": "#/components/schemas/NullableOfGenSourceType"
          },
          "genSteps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenStep"
            }
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "actionStatus": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        }
      },
...
      "GenStep": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "content": {
            "maxLength": 100000,
            "type": "string",
            "nullable": true
          },
          "outputContent": {
            "maxLength": 100000,
            "type": "string",
            "nullable": true
          },
          "path": {
            "maxLength": 400,
            "type": "string",
            "nullable": true
          },
          "outputPath": {
            "maxLength": 400,
            "type": "string",
            "nullable": true
          },
          "genStepType": {
            "$ref": "#/components/schemas/GenStepType"
          },
          "fileType": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "genActions": {
            "type": "array",
            "items": { }
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        }
      },
...
      "GenStepType": {
        "type": "integer",
        "x-enumNames": [
          "File",
          "Command",
          "Script"
        ],
        "x-enumData": [
          {
            "name": "File",
            "value": 0,
            "description": "模板生成"
          },
          {
            "name": "Command",
            "value": 1,
            "description": "执行命令"
          },
          {
            "name": "Script",
            "value": 2,
            "description": "执行脚本"
          }
        ]
      },
...
      "NullableOfGenSourceType": {
        "type": "integer",
        "nullable": true
      },
...
      "NullableOfSolutionType": {
        "type": "integer",
        "nullable": true
      },
...
      "Project": {
        "required": [
          "name",
          "displayName",
          "path"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "displayName": {
            "maxLength": 100,
            "type": "string"
          },
          "path": {
            "maxLength": 200,
            "type": "string"
          },
          "version": {
            "maxLength": 20,
            "type": "string",
            "nullable": true
          },
          "solutionType": {
            "$ref": "#/components/schemas/NullableOfSolutionType"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "apiDocInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDocInfo"
            }
          },
          "genActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GenAction"
            }
          },
          "genSteps": {
            "type": "array",
            "items": { }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdTime": {
            "type": "string",
            "format": "date-time"
          },
          "updatedTime": {
            "type": "string",
            "format": "date-time"
          },
          "isDeleted": {
            "type": "boolean"
          }
        }
      },
      "ProjectConfig": {
        "type": "object",
        "properties": {
          "idType": {
            "type": "string"
          },
          "createdTimeName": {
            "type": "string"
          },
          "updatedTimeName": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "sharePath": {
            "type": "string"
          },
          "entityPath": {
            "type": "string"
          },
          "entityFrameworkPath": {
            "type": "string"
          },
          "commonModPath": {
            "type": "string"
          },
          "apiPath": {
            "type": "string"
          },
          "microservicePath": {
            "type": "string"
          },
          "solutionPath": {
            "type": "string"
          },
          "isLight": {
            "type": "boolean"
          },
          "controllerType": {
            "$ref": "#/components/schemas/ControllerType"
          },
          "isSplitController": {
            "type": "boolean",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}